+++ docs/README.md
@@ -1,4 +1,4 @@
-# Noah User Guide
+# noah.ui.Noah User Guide
 
 // Update the title above to match the actual product name
