+++ src/main/java/noah/command/ByeCommand.java
@@ -1,3 +1,9 @@
+package noah.command;
+
+import noah.task.TaskList;
+import noah.ui.UI;
+import noah.storage.Storage;
+
 public class ByeCommand extends Command {
 
     @Override
