+++ src/main/java/noah/command/DeadlineCommand.java
@@ -1,3 +1,12 @@
+package noah.command;
+
+import noah.task.TaskList;
+import noah.ui.UI;
+import noah.storage.Storage;
+import noah.exception.NoahException;
+import noah.task.Task;
+import noah.task.Deadline;
+
 import java.io.IOException;
 import java.time.LocalDateTime;
