+++ src/main/java/noah/exception/NoahException.java
@@ -1,3 +1,5 @@
+package noah.exception;
+
 public class NoahException extends Exception {
     public NoahException(String message) {
         super(message);
