+++ src/main/java/noah/task/Deadline.java
@@ -1,3 +1,7 @@
+package noah.task;
+
+import noah.datetime.DateTime;
+
 import java.time.LocalDateTime;
 
 public class Deadline extends Task {
